+++ CHANGELOG.md
@@ -14,6 +14,7 @@ title: Changelog
 -   `@include` and `@includeCode` now work for comments on the entry point for projects with a single entry point, #2800.
 -   Cascaded modifier tags will no longer be copied into type literals, #2802.
 -   `@summary` now works to describe functions within modules, #2803.
+-   Corrected navigation showing module link as current when not on module page, #2805.
 
 ## v0.27.3 (2024-12-04)